Tweaked my inbox rules
I put all of the rules that filter automated mail at the top of the list
and added a rule that shows a desktop alert for all incoming mail in the
hopes that this will get the behavior I desire (i.e. show an alert for
e-mails that are also being moved into a folder).