Replaced most client-only rules with web app rules
The general pattern is that all the automated e-mails that I don't
really want notifications for precede the client-only rule for desktop
notifications.
For client e-mails, the rule that checks the sender's address
immediately precedes the rule that checks the recipient's address, and
only the latter stops more rules from being applied to the message.
I also removed the implementation category assignment from most
(hopefully all) of the rules, since it hasn't benefitted me in any way.
Overall, these changes are untested, so we'll see how this plays out.