Add SSH config
~/.ssh/id_rsa and ~/.ssh/id_rsa.pub are the default names for the
private and public keys, respectively, which (apparently) means the SSH
client can find the key it needs without help as long as the defaults
are used. I set up my work laptop to have distinct keys for GitHub and
BitBucket, so I configured SSH to use the correct files for the
respective hostnames and stowed it.

The keys for GitHub have a different name because GitHub is encouraging
a different keygen option (newer? better?). It's not supported on all
systems, so if you are on a system that doesn't support it, think
carefully before mucking up this file.