Fix tmux escape delay
Short-ish version: <Esc> delay in Vim was unbearable in tmux on Ubuntu
22.04 LTS in WSL, and every time I started or attached to a tmux
session, "0;10;1c" would be printed in the terminal I attached to.
Basically, this is because Ubuntu 20.04+ don't play nicely with the line
in my tmux config that sets the <Esc> delay to 0 ms. Setting it to 1 ms
fixed the issue.

Source: https://github.com/microsoft/WSL/issues/5931#issuecomment-1296783606