[bashrc] Modified tmux alias
Now, if the first positional argument to the alias is a valid directory,
it is used as the starting directory for all of the sessions and is set
as the working directory for the target session when attaching.

Fingers crossed this doesn't have some terrifying edge case I didn't
consider. It seemed to work well enough on the happy paths.