[README] Add note about git workaround
Previous commit added a workaround for git in WSL1 when accessing the
Windows file system. Because the workaround is just to leverage git on
Windows while in the Windows file system, git configurations on WSL
(obviously) don't apply. At time of writing, my shim is just to copy
over git's dotfiles, since I don't change them /that/ often. I'll find a
better fix as and when this becomes a bigger issue.