[nvim] Small reorg
I /was/ trying to keep the plugins in alphabetical order based on my
comment groupings, but I'm /more/ interested in, to the degree that it's
sensible, keeping plugin settings in the same relative order as their
corresponding 'Plug' statements. Which uhhhhh might not actually be that
helpful, now that I think about it? I don't know. It made sense at the
time.