[nvim] Add fzf actions
I guess I have to re-create the ENTIRE dictionary if I don't want to
overwrite the defaults...but the whole point was to create a mapping for
sending the selections to the quickfix window without having to open a
new view. In that vein, TIL that Alt + a selects all the results!