[nvim] Remove csharp_ls init_options
It seems like csharp_ls automatically starts in response to `:e` on a C#
file when it's been run in the session and stopped...so nothing to do
with this setting?

This whole thing is so obnoxious. I haven't been able to find any decent
documentation on the options in `omnisharp.json`, so it's anybody's
guess what any of this crap actually does. Same for a lot of the
language server settings. Very frustrating. Maybe I just need to read
the LSP docs.