[reorg-rhel] Show inactive windows' numbers
I cannot, for the life of me, figure out how to get separators to work
in inactive_winbar. Whatever.

Seems to me there might be some kind of vim variable that contains a
given window's number, but in the absence of that, this shows each
inactive window's number in its winbar, in a right-side section that
doesn't use primary colors. Looks pretty good to me.