[reorg-rhel] Add some snippets for LuaSnip
I just had to spend a bunch of time piecing a bunch of things together
to get this working. I'm still not 100% sure what each of the things I
had to add are doing to contribute to the solution, but I am maybe 75%
certain that everything I added is necessary.