Add fd to Windows
I'm not totally sure what it's bringing to the table / whether/how it's
going to help me.

Side note, just so this is written somewhere (and so I might be more
likely to remember it on account of having typed it out once):
`telescope.fd` is allegedly an alias for `telescope.find_files`, but I
have `hidden = true` for `find_files`, and hidden files show up in
`find_files` but not in `fd`. /I/ like this, I find it convenient, but I
don't know whether this is intended behavior (and therefore liable to
change at some point).