Update telescope "leader" key
I forgot that <C-M> is a carriage return. That means that if I don't
finish the rest of the key map quickly enough, telescope just receives a
carriage return, which typically defaults to doing something with the
highlighted result. Not ideal.

This simplifies the process of changing the telescope "leader" key (that
I made up) and changes it to <C-,>, which looks like it'll just insert a
comma if it times out waiting for further disambiguation. As an added
bonus, I'm guessing most/all terminals don't have some arcane meaning
assigned to <C-,>, so there shouldn't be any issues.