[windows] Fix git SSH situation on Windows
Sources:
- https://superuser.com/a/1472171
- https://superuser.com/questions/1776042/windows11-ssh-agent-does-not-remember-passphrase

vim-fugitive was giving me a "publickey denied" error, but git in
PowerShell was just prompting me for the password to the SSH key. I
think vim-fugitive just can't handle password prompts? I'm not sure
whether the SSH configs were necessary, so here they were, just in case:
In `.ssh/config`:
```
Host *
    AddKeysToAgent yes
    IdentitiesOnly yes
```

It was probably necessary to set the OpenSSH Authentication Agent to
auto-start and to run `ssh-add ~/.ssh/github...` to add the private key
to it (which should only be necessary once? I guess it just remembers
from then on out).